+++ Day4.cs
@@ -13,9 +13,9 @@ namespace AdventOfCode2020
 
         protected override List<string> ProcessInput(List<string> rawInput) => Utils.ConcatGroupOfLines(rawInput, " ", string.Empty);
 
-        public override int Part1() => CountValidPassports((_, __) => true);
+        public override long Part1() => CountValidPassports((_, __) => true);
 
-        public override int Part2()
+        public override long Part2()
         {
             return CountValidPassports((key, value) =>
             {
